Add annotations for API visibility/stability
API visibility:

- Private: only for internal usage
- VisibleForTesting: only for testing purposes
- Advanced: for usafe code to use by advance client-code

API stability:

- Experimental: might be changed/removed in future releases.
- Beta: might be changed in future releases, but not removed.

In addition, it relocates the ParamUtils class to an internal
package, to mark the code as htjdk-only usage.

The information on how to use the annotations (mostly on the
javadoc) and how to name internal/private packages should be
defined in developer guidelines checked in the repository,
but are out of the scope of this commit.